web app/requirements.txt
pandas
opencv-python
numpy
flask
pillow
werkzeug
requests
